+++ angular/projects/wcs-angular/src/lib/proxies.ts
@@ -887,13 +887,13 @@ export declare interface WcsModal extends Components.WcsModal {
 
 @ProxyCmp({
   defineCustomElementFn: undefined,
-  inputs: ['show', 'showCloseButton', 'withoutBackdrop']
+  inputs: ['show', 'showCloseButton', 'size', 'withoutBackdrop']
 })
 @Component({
   selector: 'wcs-modal',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: '<ng-content></ng-content>',
-  inputs: ['show', 'showCloseButton', 'withoutBackdrop']
+  inputs: ['show', 'showCloseButton', 'size', 'withoutBackdrop']
 })
 export class WcsModal {
   protected el: HTMLElement;