+++ src/components/modal/modal-interface.ts
@@ -1,0 +1,1 @@
+export type ModalSize = 's' | 'm' | 'l' | 'xl';