+++ src/components/modal/modal.tsx
@@ -1,4 +1,5 @@
 import { Component, Event, EventEmitter, h, Host, Listen, Prop } from '@stencil/core';
+import { ModalSize } from './modal-interface';
 
 @Component({
     tag: 'wcs-modal',
@@ -9,12 +10,12 @@ export class Modal {
     /**
      * Specifies whether the component should display a backdrop on the entire page
      */
-    @Prop({ reflect: true }) withoutBackdrop: boolean = false;
+    @Prop({reflect: true}) withoutBackdrop: boolean = false;
 
     /**
      * Displays the modal
      */
-    @Prop({ reflect: true }) show: boolean = false;
+    @Prop({reflect: true}) show: boolean = false;
 
     /**
      * Triggered when the user leaves the dialog with the closing button.
@@ -25,19 +26,25 @@ export class Modal {
      * Specifies whether the component should display a close button.
      * if false, it won't close the modal when the escape key is pressed.
      */
-    @Prop({ reflect: true }) showCloseButton: boolean = false;
+    @Prop({reflect: true}) showCloseButton: boolean = false;
+
+    /**
+     * There are multiple sizes for modals. The default size is medium (m), however other sizes are available. Select the
+     * size best suited for the content and screen size displaying the modal. Remember to test responsiveness.
+     */
+    @Prop() size: ModalSize = 'm';
 
     render() {
         return (
             <Host>
                 <div class="wcs-modal-backdrop"></div>
-                <div class="wcs-modal-container">
+                <div class="wcs-modal-container" data-size={this.size}>
                     <div class="wcs-modal-header">
                         <h5>
                             <slot name="header"></slot>
                         </h5>
                         {this.showCloseButton && (
-                            <wcs-button shape="round" mode="stroked" class="wcs-dark"
+                            <wcs-button shape="round" mode="clear" class="wcs-dark"
                                         onClick={($event) => this.onCloseButtonClick($event)}>
                                 <i class="material-icons">close</i>
                             </wcs-button>)
@@ -55,7 +62,7 @@ export class Modal {
         );
     }
 
-    @Listen('keydown', { target: 'document' })
+    @Listen('keydown', {target: 'document'})
     // @ts-ignore
     private onKeyDown(event: KeyboardEvent) {
         if (this.showCloseButton && event.key === 'Escape') {