+++ stories/modal/modal.stories.ts
@@ -1,5 +1,8 @@
 import { Meta, Story } from '@storybook/web-components';
 import { html } from 'lit-html';
+import { ifDefined } from 'lit-html/directives/if-defined';
+import { createRef, ref } from 'lit-html/directives/ref';
+import { ModalSize } from '../../src/components/modal/modal-interface';
 
 export default {
     title: 'Example/Modal',
@@ -7,28 +10,45 @@ export default {
     parameters: {
         actions: {
             handles: [
-                'wcsDialogClosed'
+                'wcsDialogClosed',
             ]
         }
     }
 } as Meta;
 
-const Template: Story<Partial<{ withoutBackdrop: boolean, show: boolean, showCloseButton: boolean }>> = (args) => html`
-    <wcs-modal ?show=${args.show}
-               ?without-backdrop=${args.withoutBackdrop}
-               ?show-close-button=${args.showCloseButton}>
-        <div slot="header">Titre de la modale</div>
-        Voulez-vous quittez la page ?
-        <p>
-            Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a cursus mi. Nullam et sem mi. Interdum et
-            malesuada fames ac ante ipsum primis in faucibus. Fusce sollicitudin pellentesque libero nec elementum.
-        </p>
-        <div slot="actions">
-            <wcs-button @click="${_ => alert('Annuler')}" mode="stroked">Annuler</wcs-button>
-            <wcs-button @click="${_ => alert('OK')}">OK</wcs-button>
-        </div>
-    </wcs-modal>
-`;
+const Template: Story<Partial<{ withoutBackdrop: boolean, show: boolean, showCloseButton: boolean, size: ModalSize }>> = (args) => {
+    // FIXME
+    // @ts-ignore
+    const modalRef = createRef<HTMLWcsModalElement>();
+    return html`
+        <wcs-button @click="${_ => {
+            modalRef.value.show = true;
+        }}">Show Modal ${args.size ? '(size: ' + args.size + ')' : null}
+        </wcs-button>
+        <wcs-modal ?show=${args.show}
+                   ${ref(modalRef)}
+                   ?without-backdrop=${args.withoutBackdrop}
+                   ?show-close-button=${args.showCloseButton}
+                   size=${ifDefined(args.size)}>
+            <div slot="header">Titre de la modale</div>
+            Voulez-vous quittez la page ?
+            <p>
+                Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a cursus mi. Nullam et sem mi. Interdum et
+                malesuada fames ac ante ipsum primis in faucibus. Fusce sollicitudin pellentesque libero nec elementum.
+            </p>
+            <div slot="actions">
+                <wcs-button @click="${_ => {
+                    modalRef.value.show = false;
+                }}" mode="stroked">Annuler
+                </wcs-button>
+                <wcs-button @click="${_ => {
+                    modalRef.value.show = false;
+                }}">OK
+                </wcs-button>
+            </div>
+        </wcs-modal>
+    `;
+};
 
 export const Default = Template.bind({});
 Default.args = {
@@ -62,3 +82,18 @@ const TemplateWithoutActions: Story<Partial<{}>> = (_) => html`
 
 export const WithoutActions = TemplateWithoutActions.bind({});
 WithoutActions.args = {};
+
+const SizeTemplate: Story<Partial<{ withoutBackdrop: boolean, show: boolean, showCloseButton: boolean, size: ModalSize }>> = (args) => html`
+    <p>For this story, the size param is not configurable, it is set manually for each button to easily show all
+        the different available sizes.</p>
+    ${Template({...args, show: false, size: 's'})}
+    ${Template({...args, show: false, size: 'm'})}
+    ${Template({...args, show: false, size: 'l'})}
+    ${Template({...args, show: false, size: 'xl'})}
+`;
+
+export const Sizes = SizeTemplate.bind({});
+Sizes.args = {
+    show: true,
+    showCloseButton: true,
+};